backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pymysql==1.1.0
cryptography==42.0.7
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
joblib==1.4.2
python-dotenv==1.0.1
apscheduler==3.10.4
bcrypt==3.2.2
